requirements.txt
openai
logging
nest_asyncio
telegram
pillow
requests
SpeechRecognition
pydub